Added a really dumb way to copy parameter values.
We're not in a secure context, apparently, so navigator.clipboard
(which seems to be the modern way of accomplishing what I want)
doesn't seem to work. The simplest workaround I found for this is
just using the prompt() function and insructing the user to copy
the values from there.

Also I had to do a weird function...closure? I don't know if what
I did counts as a closure. But there was some weirdness going on
with an incorrect 'this' context where I feel like there shouldn't
be...