Added functionality to keep track of which data is currently being plotted. Can now plot any CL from values. Still want to change things so re-running PSO will plot the currently selected CL (unless that data's been removed.)
There's also a weird error I think is unrelated to what I've done
in this branch: load data, run PSO, then load more data and click
"plot" on it, and it errors our. Gotta see what that's about.
Should be easy to fix, just hoping it's not awful to diagnose.