Get the 4v Brugada code working with sample data
The main issue was that the wrong parameter bounds were being used. The
issue was fixed by adjusting the bounds, although the fits still aren't
as good as when certain parameters are fixed as in the reference
implementation. Right now, the fixed versions are used and the bounds
for those parameters are commented out.

One significant difference between the WebGL implementation and the
reference implementation remains, which is that the reference version
scales the chi parameter by 0.25 from its standard value. I tested both
and it doesn't seem to make much of a difference in the quality; since
chi is a hyperparameter that can be adjusted in the UI it certainly
doesn't impact the "legitimacy" of the results.

Also, the reference implementation and data are now included in the repo
to save everyone some trouble.