Add logging of convergence data
Set the "dump_convergence" variable to true to save particle states
after each iteration. This information is used to track the convergence
of the algorithm, but comes at a performance cost and should generally
not be turned on when not needed.

Additionally, these changes include improved hyperparameter values
based on convergence results and the ability to choose a larger
particle count.